Make typing relation allow cyclic data
Since variables point to multiequations that contain sets of those same
variables, we will clearly need cyclic data.
The typing relation is extended with a set of "cached" addresses for
which the type is unconditionally inferred, and an address is added to the
cache during its derivation. (The cache should start empty.)

Fixed proofs as necessary. The old theorem that a well-typed bound value
cannot reach its address is now false: reaching your own address means
having a cycle. However, the reachability condition for type
preservation when the store is updated is no longer required.
(Reachability relations will probably not be so important any more.)