Defined first half of reduce; includes approach to loop variables
May not be correct, but at least it type checks.

In the body of a loop, we augment the state with an extra piece of "loop
state" to store any local loop variables. loop_lift lifts things in the
underlying monad to pass the loop state through, and we add an overload
so we don't have to write loop_lift everywhere. loop_get and loop_put
get and put the contents of the loop state, hopefully simulating the
reading and writing of real lexical variables.