Make while loops return the loop state (instead of unit)
Necessary for when assignments to lexical variables within a loop are
relevant after the loop.

Amounts to making the loop state projection function a lossless
projection.

Also, made the definitions of raw_while and raw_repeat, saving while and
repeat for overloads of the common case: injection is ($, t) and
projection is simply I.