Fix congruence for foo, but another failing example
After stepping through the TFL code with me, Michael figured out a form
of the congruence theorem for STATE_OPTION_BIND that would allow useful
termination conditions to be extracted for "foo".

This commit includes a slightly more complicated definition, "plen",
which doesn't succumb to the same trick. I think the congruence theorem
that works for foo is too weak, leading to unprovable termination
conditions for plen.