Redesign semi-automatic way to get good TCs
Previous commit was broken: termination goal for plen was probably not
provable. Here we try a different approach. Instead of inverse-eta
expanding certain monadic constants, we instead make sure that
applications of BIND and IGNORE_BIND have the right number of arguments,
and that the second argument is an abstraction with at least 2 (or 1)
bound variables.

This creates almost all the right matching conditions for TFL to operate
nicely, but there is (at least) one ugly hack required, because the
conversion can introduce terms of the form ((if b then a else c) s). The
hack is to rewrite these to (if b then a s else c s) after the
conversion, before passing off to TFL. I don't know if there may be more
cases like this that need a rewrite; probably not in any code I will
write.